refactor : Colors enum 위치 수정
Piece 내부로 enum의 위치를 수정, 수정에 따른 테스트 코드 수정
