--- src/chess/pieces/Colors.java
@@ -1,16 +1,0 @@
-package src.chess.pieces;
-
-public enum Colors {
-    WHITE("white"),
-    BLACK("black");
-
-    private final String colorName;
-
-    Colors(String colorName) {
-        this.colorName = colorName;
-    }
-
-    public String getColorName() {
-        return colorName;
-    }
-}
