+++ src/test/pieces/PieceTest.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.DisplayName;
-import src.chess.pieces.Colors;
 import src.chess.pieces.Piece;
 import org.junit.jupiter.api.Test;
 
@@ -12,8 +11,8 @@ public class PieceTest {
     @Test
     @DisplayName("폰을 생성할 때 올바른 색의 폰인지, 색에 따라 올바른 기물 모양이 세팅되었는지 확인한다.")
     public void create_piece() {
-        verifyPiece(Piece.createWhitePawn(), Colors.WHITE.getColorName(), Piece.Type.PAWN.getPieceByColor(Colors.WHITE));
-        verifyPiece(Piece.createBlackPawn(), Colors.BLACK.getColorName(), Piece.Type.PAWN.getPieceByColor(Colors.BLACK));
+        verifyPiece(Piece.createWhitePawn(), Piece.Colors.WHITE.getColorName(), Piece.Type.PAWN.getPieceByColor(Piece.Colors.WHITE));
+        verifyPiece(Piece.createBlackPawn(), Piece.Colors.BLACK.getColorName(), Piece.Type.PAWN.getPieceByColor(Piece.Colors.BLACK));
     }
 
     private void verifyPiece(final Piece piece, final String color, final String representation) {
@@ -40,8 +39,8 @@ public class PieceTest {
     @Test
     @DisplayName("폰의 기물 모양이 색에 맞게 올바르게 관리되고 있는지 확인한다.")
     public void getRepresentationPerPiece() throws Exception {
-        assertThat("p").isEqualTo(Piece.Type.PAWN.getPieceByColor(Colors.WHITE));
-        assertThat("P").isEqualTo(Piece.Type.PAWN.getPieceByColor(Colors.BLACK));
+        assertThat("p").isEqualTo(Piece.Type.PAWN.getPieceByColor(Piece.Colors.WHITE));
+        assertThat("P").isEqualTo(Piece.Type.PAWN.getPieceByColor(Piece.Colors.BLACK));
     }
 }
